External harness (#7)
* Switch to external harness

The rock tests are expected to share as significant amount of code
(e.g. k8s harness, various helpers), which was moved to a separate
project to avoid duplication.

While at it, we're fixing the linter and code formatter jobs.

* Update integration tests, failing early if no image is provided

We'll ensure that the integration tests fail early if the rock
images are not specified.