Fixes integration tests (#11)
The bitnami helm chart requires the image registry to be separated from
the image name. Without this, it would try to use something like the
following as an image name: docker.io/ghcr.io/repository/multus:tag,
which is incorrect.

The integration tests currently work if the rock image is published on
docker.io, but not on ghcr.io.